% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/No-Cor-flashfm-v20240222_update.R
\name{calcAdjPP0}
\alias{calcAdjPP0}
\title{Calculates trait-adjusted posterior probabilities for all traits at sharing parameter kappa}
\usage{
calcAdjPP0(qt, STR, SS, tau, nsnpspermodel, kappa, PP, beta, NCORES)
}
\arguments{
\item{qt}{vector of trait names}

\item{STR}{list consisting of vectors of model configurations for each trait}

\item{SS}{list consisting of lists of model configuration SNPs for each trait}

\item{tau}{matrix of adjustment terms}

\item{nsnpspermodel}{list of number of SNPs per model for each model in STR}

\item{kappa}{single value of sharing parameter kappa}

\item{PP}{list consisting of vectors of posterior probabilities for the model configurations for each trait}

\item{beta}{list of joint effect estimates for models in STR; multi.beta output}

\item{NCORES}{number of cores for parallel computing; recommend NCORES=M, but if on Windows, use NCORES=1;}
}
\value{
list of trait-adjusted posterior probabilities for each trait at sharing parameter kappa
}
\description{
Calculates trait-adjusted posterior probabilities for all traits at sharing parameter kappa
}
\author{
Jenn Asimit
}
