% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiddenFunctions.R
\name{.ReweightSnpEffectsAccordingToBlockCorrelations}
\alias{.ReweightSnpEffectsAccordingToBlockCorrelations}
\title{This function is used by JAMPred_Step2AdjustmentAndPredictions}
\usage{
.ReweightSnpEffectsAccordingToBlockCorrelations(iteration, snp.blocks, X.ref)
}
\arguments{
\item{iteration}{The number of iteration}

\item{snp.blocks}{The blocks of SNPs}

\item{X.ref}{X ref}
}
\description{
This function is used by JAMPred_Step2AdjustmentAndPredictions
}
