% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentGWAS.R
\name{factor_contributions}
\alias{factor_contributions}
\title{Re-scale factor loadings to get their contributions and output factor loadings and contributions matrices with observed traits ordered by maximum contributing latent factor}
\usage{
factor_contributions(L)
}
\arguments{
\item{L}{PxK matrix of factor loadings generated from a factor analysis of P observed traits and K latent factors (K<P)}
}
\value{
factor loadings and re-scaled factor loadings (contributions) matrices with observed traits ordered by maximum contributing latent factor
}
\description{
Re-scale factor loadings to get their contributions and output factor loadings and contributions matrices with observed traits ordered by maximum contributing latent factor
}
\author{
Jenn Asimit
}
