% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faYloading.R
\docType{data}
\name{faYloading}
\alias{faYloading}
\title{Factor loading matrix for four latent factors estimated from 10 blood cell traits of
the INTERVAL data, consisting of 43,059 participants.
For the factor analysis, the covariance matrix was calculated using pair-wise complete data,
Original INTERVAL GWAS data are available in \href{https://doi.org/10.1016/j.cell.2016.10.042}{Astle et al. (2016)}
and \href{https://doi.org/10.1038/s41467-023-40679-y}{Akbari et al. (2023)}}
\format{
An object of class \code{loadings} with 10 rows and 4 columns.
}
\usage{
faYloading
}
\description{
Factor loading matrix for four latent factors estimated from 10 blood cell traits of
the INTERVAL data, consisting of 43,059 participants.
For the factor analysis, the covariance matrix was calculated using pair-wise complete data,
Original INTERVAL GWAS data are available in \href{https://doi.org/10.1016/j.cell.2016.10.042}{Astle et al. (2016)}
and \href{https://doi.org/10.1038/s41467-023-40679-y}{Akbari et al. (2023)}
}
\keyword{datasets}
