% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/No-Cor-flashfm-v20240222_update.R
\name{flashfmZero.input}
\alias{flashfmZero.input}
\title{Key input for flashfm - constructs snpmod object list and joint effect estimates list for all trait if have external single trait fine-mapping results, when trait correlation is zero}
\usage{
flashfmZero.input(modPP.list, beta1.list, corX, Nall, ybar.all, raf)
}
\arguments{
\item{modPP.list}{list of data.frame objects for each trait, containing a column named "snps": snp models of the form \code{"snp1,snp2"} and "PP": snp model posterior probability from single trait fine-mapping}

\item{beta1.list}{list of single SNP effect estimates for each trait in the form of a named vector; name of each effect estimate should appear in Gmat}

\item{corX}{genotype matrix with SNP columns; could be from sample or reference panel; use unrelated samples}

\item{Nall}{vector of sample sizes; if related samples then give effective sample sizes}

\item{ybar.all}{vector of trait means; if related samples, this should be based on unrelated samples; if traits are transformed to be standard Normal, could set ybar as 0-vector}

\item{raf}{named vector of SNP reference allele frequencies where name is snp id (must match SNP coding used for effect estimates); only needed if Gmat is a covariance matrix and MUST be in same order as SNPs in covariance matrix}
}
\value{
list containing the main input for flashfm
}
\description{
Key input for flashfm - constructs snpmod object list and joint effect estimates list for all trait if have external single trait fine-mapping results, when trait correlation is zero
}
\author{
Jenn Asimit
}
