% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/No-Cor-flashfm-v20240222_update.R
\name{flashfmZero}
\alias{flashfmZero}
\title{Marginal PP for models of a set of traits, sharing information between the traits, when trait correlation is zero}
\usage{
flashfmZero(main.input, TOdds = 1, cpp = 0.99, maxmod = NULL, NCORES)
}
\arguments{
\item{main.input}{List of 3 components: SM=list of snpmod objects for a set of traits; mbeta=list of joint effects for each trait; nsnps= number of SNPs in the region
This could be obtained from flashfm.input or JAMexpanded.multi.}

\item{TOdds}{Vector of target odds of no sharing to sharing}

\item{cpp}{cumulative posterior probability threshold for selecting top models; this is ignored when maxmod is spespecified}

\item{maxmod}{maximum number of top models to output; NULL by default}

\item{NCORES}{number of cores for parallel computing; set NCORES=M (number of traits) or NCORES=1 if running on hpc (do not set an arbitrary number of cores), but if on Windows/Mac, must use NCORES=1}
}
\value{
List consisting of PP: marginal PP for models and MPP: marginal PP of SNP inclusion
}
\description{
Marginal PP for models of a set of traits, sharing information between the traits, when trait correlation is zero
}
\author{
Jenn Asimit
}
