% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table_FMsummaries.R
\name{FMsummary_table}
\alias{FMsummary_table}
\title{Make table of fine-mapping results output from "flashfm" (flashfmZero or flashfm), "multiJAMd", or "JAMdwithGroups"}
\usage{
FMsummary_table(FMout, method = "flashfm", traitnames = NULL, cred = 0.99)
}
\arguments{
\item{FMout}{object output from "flashfm" (flashfmZero or flashfm related wrappers, like "FLASHFMZEROwithJAMd"), "multiJAMd", or "JAMdwithGroups"}

\item{method}{text for fine-mapping method used, i.e. one of "flashfm" (flashfmZero or flashfm), "multiJAMd", or "JAMdwithGroups"}

\item{traitnames}{vector of trait names (same length as number of traits that were fine-mapped) and in same order as fine-mapped traits}

\item{cred}{level used for credible set construction; default is 0.99}
}
\value{
Table where each row corresponds to one trait, listing credible set size, details of variant with maximum MPP, variants with MPP>0.90 and snp groups coinciding with the variant(s)
}
\description{
Make table of fine-mapping results output from "flashfm" (flashfmZero or flashfm), "multiJAMd", or "JAMdwithGroups"
}
\author{
Jenn Asimit
}
