% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makegroups1trait_v20240222_update.R
\name{JAMdwithGroups}
\alias{JAMdwithGroups}
\title{Wrapper for JAMdynamic single-trait fine-mapping that also outputs SNP groups; PP and credible sets include SNP group information}
\usage{
JAMdwithGroups(
  gwas,
  N,
  corX,
  save.path,
  cred = 0.99,
  jam.nM.iter = 1,
  maxcv = 1,
  maxcv_stop = 20,
  min.mppi = 0.01,
  minsnpmppi = 0.01,
  r2.minmerge = 0.6,
  extra.java.arguments = NULL
)
}
\arguments{
\item{gwas}{data.frame with 3 columns named: rsID, beta, EAF;}

\item{N}{scalar that is the trait sample size}

\item{corX}{SNP correlation matrix that must have snp names in row.names and col.names}

\item{save.path}{Path to save JAM output files; tmp files and could delete these later e.g. save.path=paste0(DIRout,"/tmpJAM/region1").}

\item{cred}{probability for credible set; default is 0.99}

\item{jam.nM.iter}{in millions, number of iterations to use in JAM; default 1 (1 million)}

\item{maxcv}{starting value for maximum number of causal variants}

\item{maxcv_stop}{maximum value to consider for maximum number of causal variants; maxcv_stop >= maxcv.}

\item{min.mppi}{trim snp groups with total MPPI < min.mppi in all diseases; default 0.01}

\item{minsnpmppi}{only group snps with total MPPI > minsnpmppi; default 0.01}

\item{r2.minmerge}{merge groups with minimum between-group r2 > r2.minmerge; default 0.5}

\item{extra.java.arguments}{A character string to be passed through to the java command line. E.g. to specify a
different temporary directory by passing "-Djava.io.tmpdir=/Temp".}
}
\value{
List consisting of three objects: CS, a data.frame detailing the SNPs in the credible set; mpp.pp, a list with 4 components giving the SNP-level results (mpp.pp$PP,mpp.pp$MPP) and SNP group level results (mpp.pp$MPPg, mpp.pp$PPg); and snpGroups,
a list giving the SNP groups construced under single-trait fine-mapping
}
\description{
Wrapper for JAMdynamic single-trait fine-mapping that also outputs SNP groups; PP and credible sets include SNP group information
}
\author{
Jenn Asimit
}
