% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentGWAS.R
\name{latentBeta}
\alias{latentBeta}
\title{GWAS effect estimates for latent factors at one variant using observed trait GWAS summary statistics}
\usage{
latentBeta(beta_traits, Lterm)
}
\arguments{
\item{beta_traits}{vector of effect estimates from a set of observed trait GWAS}

\item{Lterm}{loadingterm matrix output from LoadingTerm function}
}
\value{
vector of effect estimates for the latent factors (at one variant)
}
\description{
GWAS effect estimates for latent factors at one variant using observed trait GWAS summary statistics
}
\author{
Jenn Asimit
}
