% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentGWAS.R
\name{latentBetaSE}
\alias{latentBetaSE}
\title{GWAS effect estimate standard errors for latent factors at one variant using observed trait GWAS summary statistics}
\usage{
latentBetaSE(se_traits, Lterm, covY)
}
\arguments{
\item{se_traits}{vector of standard errors}

\item{Lterm}{loadingterm matrix output from LoadingTerm function}

\item{covY}{trait covariance matrix (same as used in factor analysis)}
}
\value{
vector of effect estimate standard errors for the latent factors (at one variant)
}
\description{
GWAS effect estimate standard errors for latent factors at one variant using observed trait GWAS summary statistics
}
\author{
Jenn Asimit
}
