% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentGWAS.R
\name{latentGWAS}
\alias{latentGWAS}
\title{GWAS for latent factors at one variant using observed trait GWAS summary statistics}
\usage{
latentGWAS(
  obsGWAS,
  covY,
  L,
  beta_colname = "BETA",
  se_colname = "SE",
  snpID_colname = "SNP",
  EAF
)
}
\arguments{
\item{obsGWAS}{list of observed GWAS data.frames, containing the effect estimates and standard errors and snp names, at a minimum}

\item{covY}{trait covariance matrix (same as used in factor analysis)}

\item{L}{PxK matrix of factor loadings generated from a factor analysis of P observed traits and K latent factors (K<P)}

\item{beta_colname}{text column name for the effect estimates in each obsGWAS data.frame; default "BETA"}

\item{se_colname}{text column name for the effect estimate standard errors in each obsGWAS data.frame; default "SE"}

\item{snpID_colname}{text column name for the effect estimates in each obsGWAS data.frame; default "SNP"}

\item{EAF}{vector of effect allele frequencies in same order as variants in obsGWAS}
}
\value{
list of latent GWAS results in format for flashfmZero, with columns rsID, beta, EAF, beta_se, p-value
}
\description{
GWAS for latent factors at one variant using observed trait GWAS summary statistics
}
\author{
Jenn Asimit
}
