% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentGWAS.R
\name{LoadingTerm}
\alias{LoadingTerm}
\title{GWAS for latent factors at one variant using observed trait GWAS summary statistics}
\usage{
LoadingTerm(L)
}
\arguments{
\item{L}{PxK matrix of factor loadings generated from a factor analysis of P observed traits and K latent factors (K<P)}
}
\value{
function of factor loadings (L^T L)^{-1}L^T; needed to estimate latent GWAS effect estimates from observed traits
}
\description{
GWAS for latent factors at one variant using observed trait GWAS summary statistics
}
\author{
Jenn Asimit
}
