% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/No-Cor-flashfm-v20240222_update.R
\name{marginalpp0}
\alias{marginalpp0}
\title{Marginal PP for models sharing information between traits, when trait correlation is zero}
\usage{
marginalpp0(STR, PP, mbeta, kappa, N, nsnps, NCORES)
}
\arguments{
\item{STR}{list of models for traits 1, 2, ..., n, each given in
the form of a character vector, with entries
\code{"snp1\%snp2\%snp3"}. The null model is given by
\code{"1"} OR \code{"0"}.  It is assumed that all elements of
ABF, PP and pr below follow this same order.}

\item{PP}{list of posterir probabilities for the models in M}

\item{mbeta}{list of joint beta estimates for each trait}

\item{kappa}{single value or vector of values to consider for the
sharing scale parameter.  the value of kappa=1 must be
included, and if not will be prepended.}

\item{N}{number of individiduals with measurements for all traits}

\item{nsnps}{number of snps in region}

\item{NCORES}{number of cores for parallel computing; recommend NCORES=M, but if on Windows, use NCORES=1}
}
\value{
list of: - single.pp: list of pp for each model in
\code{STR[[i]]} for trait i - shared.pp: list of pp for each model
in \code{STR[[i]]} for trait i
}
\description{
Marginal PP for models sharing information between traits, when trait correlation is zero
}
