% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makegroups1trait_v20240222_update.R
\name{multiJAMdCS}
\alias{multiJAMdCS}
\title{Return credible sets from single-trait fine-mapping of several traits}
\usage{
multiJAMdCS(multiJAMout, cred = 0.99)
}
\arguments{
\item{multiJAMout}{output from the function multiJAMd}

\item{cred}{probability for credible set; default is 0.99.}
}
\value{
List of two items: fm is a list of single-trait credible sets for several traits; cred is the probability for constructing the credible sets.
}
\description{
Return credible sets from single-trait fine-mapping of several traits
}
\author{
Jenn Asimit
}
