% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassMethods.R
\name{R2BGLiMS_Results-methods}
\alias{R2BGLiMS_Results-methods}
\alias{show,R2BGLiMS_Results-method}
\title{'show' method for R2BGLiMS-R2BGLiMS_Results-class objects}
\usage{
\S4method{show}{R2BGLiMS_Results}(object)
}
\arguments{
\item{object}{Check NULL values.}
}
\value{
Prints a summary of the R2BGLiMS-R2BGLiMS_Results-class object.
}
\description{
Prints a summary of the R2BGLiMS-R2BGLiMS_Results-class object.
}
\author{
Paul J. Newcombe \email{paul.newcombe@mrc-bsu.cam.ac.uk}
}
