% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sel_obsgwas.R
\docType{data}
\name{sel_obsgwas}
\alias{sel_obsgwas}
\title{List of GWAS for 10 blood cell traits from the INTERVAL data, consisting of 43,059 participants.
Covariance was calculated using pair-wise complete data, i.e. some individuals have missing trait measurements.
Original INTERVAL GWAS data are available in \href{https://doi.org/10.1016/j.cell.2016.10.042}{Astle et al. (2016)}
and \href{https://doi.org/10.1038/s41467-023-40679-y}{Akbari et al. (2023)}}
\format{
An object of class \code{list} of length 10.
}
\usage{
sel_obsgwas
}
\description{
List of GWAS for 10 blood cell traits from the INTERVAL data, consisting of 43,059 participants.
Covariance was calculated using pair-wise complete data, i.e. some individuals have missing trait measurements.
Original INTERVAL GWAS data are available in \href{https://doi.org/10.1016/j.cell.2016.10.042}{Astle et al. (2016)}
and \href{https://doi.org/10.1038/s41467-023-40679-y}{Akbari et al. (2023)}
}
\keyword{datasets}
