% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMIM1.corX.R
\docType{data}
\name{SMIM1.corX}
\alias{SMIM1.corX}
\title{SNP correlation matrix for the subset of the INTERVAL data that has no missing trait measurements
Original INTERVAL GWAS data are available in \href{https://doi.org/10.1016/j.cell.2016.10.042}{Astle et al. (2016)}
and \href{https://doi.org/10.1038/s41467-023-40679-y}{Akbari et al. (2023)}}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 313 rows and 313 columns.
}
\usage{
SMIM1.corX
}
\description{
SNP correlation matrix for the subset of the INTERVAL data that has no missing trait measurements
Original INTERVAL GWAS data are available in \href{https://doi.org/10.1016/j.cell.2016.10.042}{Astle et al. (2016)}
and \href{https://doi.org/10.1038/s41467-023-40679-y}{Akbari et al. (2023)}
}
\keyword{datasets}
