% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMIM1.gwas.list.interval.fa25.R
\docType{data}
\name{SMIM1.gwas.list.interval.fa25}
\alias{SMIM1.gwas.list.interval.fa25}
\title{GWAS data for three latent traits from a subset of the INTERVAL data that has no missing trait measurements
This is a list of three latent trait GWAS data.frames, where each data.frame has columns rsID, beta, EAF, P_value
These latent traits (ML4, ML12, ML14) are all related to red blood cell traits and our analyses suggest that
rs1175550 is the causal variant for all three traits, which is in agreement with UKBB fine-mapping \href{https://doi.org/10.1016/j.cell.2020.08.008}{(Vuckovic et al. 2020)} previous studies \href{https://doi.org/10.1038/ng.2603}{(Cvejic et al. 2013)}
Original INTERVAL GWAS data are available in \href{https://doi.org/10.1016/j.cell.2016.10.042}{Astle et al. (2016)}
and \href{https://doi.org/10.1038/s41467-023-40679-y}{Akbari et al. (2023)}}
\format{
An object of class \code{list} of length 3.
}
\usage{
SMIM1.gwas.list.interval.fa25
}
\description{
GWAS data for three latent traits from a subset of the INTERVAL data that has no missing trait measurements
This is a list of three latent trait GWAS data.frames, where each data.frame has columns rsID, beta, EAF, P_value
These latent traits (ML4, ML12, ML14) are all related to red blood cell traits and our analyses suggest that
rs1175550 is the causal variant for all three traits, which is in agreement with UKBB fine-mapping \href{https://doi.org/10.1016/j.cell.2020.08.008}{(Vuckovic et al. 2020)} previous studies \href{https://doi.org/10.1038/ng.2603}{(Cvejic et al. 2013)}
Original INTERVAL GWAS data are available in \href{https://doi.org/10.1016/j.cell.2016.10.042}{Astle et al. (2016)}
and \href{https://doi.org/10.1038/s41467-023-40679-y}{Akbari et al. (2023)}
}
\keyword{datasets}
