% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopModels.R
\name{TopModels}
\alias{TopModels}
\title{Table of top models}
\usage{
TopModels(
  results,
  n.top.models = 20,
  covariates.to.include = NULL,
  remove.empty.cols = TRUE
)
}
\arguments{
\item{results}{A R2BGLiMS R2BGLiMS_Results-class object frum running the R2BGLiMS R2BGLiMS command.}

\item{n.top.models}{number of top models to include in the table (default 20).}

\item{covariates.to.include}{Character vector indicating a subset of covariates to calculate marginal posterior combinations for. NOTE: Not yet implemented for enumeration.}

\item{remove.empty.cols}{Check NULL values.}
}
\value{
Ordered table of best models, with corresponding posterior probabilities. If JAM
was run with multiple covariate blocks, a list of model tables is returned.
}
\description{
Generates a table of the top models from a Reversible Jump Results object.
}
\author{
Paul Newcombe
}
