% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TracePlots.R
\name{TracePlots}
\alias{TracePlots}
\title{Parameter posterior trace plots for a R2BGLiMS results object}
\usage{
TracePlots(
  results,
  vars.to.include = NULL,
  var.dictionary = NULL,
  predictors.only = FALSE,
  file = NULL,
  add.to.title = NULL,
  par.mfrow = NULL,
  add.panel.labels = FALSE
)
}
\arguments{
\item{results}{A R2BGLiMS R2BGLiMS_Results-class object frum running the R2BGLiMS R2BGLiMS command.}

\item{vars.to.include}{Check NULL values.}

\item{var.dictionary}{Character vector containing a mapping of covariate names to substitutions for the plot. The names of the elements are
the variable names as in the original analysis (i.e. as named in the results object).}

\item{predictors.only}{If true only chain plots for predictors are plotted (and intercept,
likelihood etc. are excluded). Default is FALSE.}

\item{file}{PDF path}

\item{add.to.title}{Optional character string to add to the start of each plot title}

\item{par.mfrow}{Allow passing the mfrow layout vector to par. Default par(mfrow=c(5,2))}

\item{add.panel.labels}{Add lower case letter labels to panel titles. Default FALSE.
Note: Does not work when providing a var.dictionary}
}
\value{
NA
}
\description{
Generates posterior trace plots for a R2BGLiMS results object
}
\author{
Paul Newcombe
}
