% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/No-Cor-flashfm-v20240222_update.R
\name{FLASHFMZEROwithJAMd}
\alias{FLASHFMZEROwithJAMd}
\title{Wrapper for flashfm Multi-Trait Fine-Mapping with JAM (when trait correlation is zero)  - this is the dynamic number of max causal variant version}
\usage{
FLASHFMZEROwithJAMd(
  gwas.list,
  corX,
  N,
  save.path,
  TOdds = 1,
  cpp = 0.99,
  NCORES,
  maxcv = 1,
  maxcv_stop = 20,
  jam.nM.iter = 1,
  r2.minmerge = 0.6,
  minsnpmppi = 0.01
)
}
\arguments{
\item{gwas.list}{List of M data.frame objects, where M is the number of traits; gwas.list[[i]] is a data.frame for  trait i with 3 columns named: rsID, beta, EAF}

\item{corX}{SNP correlation matrix}

\item{N}{Vector of length M; Nall[i] is the (effective) sample size for trait i}

\item{save.path}{Path to save JAM output files; tmp files and could delete these later e.g. save.path=paste0(DIRout,"/tmpJAM/region1").}

\item{TOdds}{target odds of no sharing to sharing; default is 1}

\item{cpp}{cumulative posterior probability threshold for selecting top models; default 0.99}

\item{NCORES}{number of cores for parallel computing; recommend NCORES=M, but if on Windows, use NCORES=1}

\item{maxcv}{starting value for maximum number of causal variants; default 1}

\item{maxcv_stop}{maximum value to consider for maximum number of causal variants; maxcv_stop >= maxcv.}

\item{jam.nM.iter}{in millions, number of iterations to use in JAM; defailt 1 (1 million)}

\item{r2.minmerge}{merge groups with minimum between-group r2 > r2.minmerge; default 0.6}

\item{minsnpmppi}{only group snps with total MPPI > minsnpmppi; default 0.01}
}
\value{
list with 2 components: mpp.pp, a list with 4 components giving the SNP-level results (mpp.pp$PP,mpp.pp$MPP) and SNP group level results (mpp.pp$MPPg, mpp.pp$PPg); and snpGroups,
a list with 2 components giving the SNP groups constructed under single-trait (snpGroups[[1]]) and multi-trait fine-mapping (snpGroups[[2]])
}
\description{
Wrapper for flashfm Multi-Trait Fine-Mapping with JAM (when trait correlation is zero)  - this is the dynamic number of max causal variant version
}
\author{
Jenn Asimit
}
