% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table_FMsummaries.R
\name{FMsummary_table_general}
\alias{FMsummary_table_general}
\title{Make table of fine-mapping results output for latent factors and observed traits (wrapper for FMsummary_table)}
\usage{
FMsummary_table_general(
  FM_obs,
  FM_latent,
  fm_traits_obs,
  fm_traits_latent,
  cred = 0.99,
  regions = NULL,
  array_index = NULL
)
}
\arguments{
\item{FM_obs}{observed trait fine-mapping object output from "multiJAMd"}

\item{FM_latent}{latent factor object output from "flashfm" (flashfmZero or flashfm related wrappers, like "FLASHFMZEROwithJAMd") or "JAMdwithGroups"}

\item{fm_traits_obs}{vector of observed trait names (same length as number of traits that were fine-mapped) and in same order as fine-mapped traits}

\item{fm_traits_latent}{vector of latent factor names (same length as number of factors that were fine-mapped) and in same order as fine-mapped factors}

\item{cred}{level used for credible set construction; default is 0.99}

\item{regions}{data.frame with region details, where the columns are in order: "chromosom"e, "start"", "end"}

\item{array_index}{row number of regions data.frame}
}
\value{
Table where each row corresponds to one trait, listing credible set size, details of variant with maximum MPP, variants with MPP>0.90 and snp groups coinciding with the variant(s)
}
\description{
Make table of fine-mapping results output for latent factors and observed traits (wrapper for FMsummary_table)
}
\author{
Jenn Asimit
}
