% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makegroups1trait_v20240222_update.R
\name{multiJAMd}
\alias{multiJAMd}
\title{Wrapper for JAMdynamic single-trait fine-mapping applied to a set of traits, and output SNP groups}
\usage{
multiJAMd(
  gwas.list,
  corX,
  N,
  save.path,
  maxcv = 1,
  maxcv_stop = 20,
  jam.nM.iter = 1,
  min.mppi = 0.01,
  minsnpmppi = 0.01,
  r2.minmerge = 0.6,
  NCORES = 1,
  extra.java.arguments = NULL
)
}
\arguments{
\item{gwas.list}{List of M data.frames, where M is the number of traits; gwas.list[[i]] is a data.frame (one for each trait)
with 3 columns named: rsID, beta, EAF;
if trait names are provided for the M data.frames, these trait names are given in output}

\item{corX}{SNP correlation matrix that must have snp names in row.names and col.names}

\item{N}{scalar that is the trait sample size - same for all traits; if different sample sizes across traits, use flashfm to return single and multi-trait results}

\item{save.path}{Path to save JAM output files; tmp files and could delete these later e.g. save.path=paste0(DIRout,"/tmpJAM/region1").}

\item{maxcv}{starting value for maximum number of causal variants}

\item{maxcv_stop}{maximum value to consider for maximum number of causal variants; maxcv_stop >= maxcv.}

\item{jam.nM.iter}{in millions, number of iterations to use in JAM; default 1 (1 million)}

\item{min.mppi}{trim snp groups with total MPPI < min.mppi in all diseases; default 0.01}

\item{minsnpmppi}{only group snps with total MPPI > minsnpmppi; default 0.01}

\item{r2.minmerge}{merge groups with minimum between-group r2 > r2.minmerge; default 0.6}

\item{NCORES}{number of cores for parallel computing; recommend NCORES=M if running on hpc, but if on Windows/Mac, use NCORES=1}

\item{extra.java.arguments}{A character string to be passed through to the java command line. E.g. to specify a
different temporary directory by passing "-Djava.io.tmpdir=/Temp".}
}
\value{
List consisting of two objects: mpp.pp, a list with 4 components giving the SNP-level results (mpp.pp$PP,mpp.pp$MPP) and SNP group level results (mpp.pp$MPPg, mpp.pp$PPg); and snpGroups,
a list giving the SNP groups construced under single-trait fine-mapping
}
\description{
Wrapper for JAMdynamic single-trait fine-mapping applied to a set of traits, and output SNP groups
}
\author{
Jenn Asimit
}
